게시글 상세보기 시 사용되는 지도를 카카오 지도에서 네이버 지도로 바꾸고 약속 장소의 위치를 마커로 표시.

+++ bob-friend-iOS/bob-friend-iOS/Model/DataModel/AppointmentModel.swift
@@ -10,14 +10,10 @@ import Foundation
 struct AppointmentSimpleModel: Decodable {
     let id: Int
     let title: String
-//    let content: String
     let author: User
     let amountOfComments: Int
     let totalNumberOfPeople: Int
     let currentNumberOfPeople: Int
-//    let full: Bool
-//    let sexRestriction
-//    let appointmentTime: String?
     let createdAt: String
 }
 
@@ -31,11 +27,10 @@ struct AppointmentModel: Decodable {
     let comments: [CommentModel]
     let totalNumberOfPeople: Int
     let currentNumberOfPeople: Int
-//    let full: Bool
     let restaurantName: String
     let restaurantAddress: String
-    let latitude: Float
-    let longitude: Float
+    let latitude: Double
+    let longitude: Double
     let sexRestriction: Gender
     let ageRestrictionStart: Int?
     let ageRestrictionEnd: Int?
